// apps/web/app/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "product-category".
 */
export interface ProductCategory {
  id: number;
  title: string;
  image: number | Media;
  secondaryImage?: (number | null) | Media;
  show?: boolean | null;
  url: string;
  updatedAt: string;
  createdAt: string;
}